Use 16-bit global indices in BLE packet format
Bump BLE_TX_QUEUE_SIZE to 64 and revise BLE packet layout to use a 4-byte global header (two 16-bit fields) instead of the previous 2-byte header. Update LOCAL_PACKETS_PER_GLOBAL from 29 to 32 and adjust packet construction/memcpy offsets accordingly. Change SendNotify signature to accept uint16_t global_total/global_idx and update the prototype in BSP/bsp.h. Files changed: BSP/bleprph.c, BSP/bsp.h. Rationale: support larger global counters (>255), align queue size to a power-of-two, and update packet framing to match the new header size.

+++ BSP/bsp.h
@@ -31,7 +31,7 @@ void gps_get_pps_stats(uint32_t *count, uint64_t *last_time_us);
 esp_err_t BSPInit(void);
 esp_err_t InitBlueTooth(void);
 void RunBlueToothHost(void);
-int SendNotify(uint8_t *buf, size_t len, uint8_t global_total, uint8_t global_idx);
+int SendNotify(uint8_t *buf, size_t len, uint16_t global_total, uint16_t global_idx);
 
 esp_err_t InitDataPeripheral(void);
 void RunDataPeripheral(void *pvParameters);
